chore(core): extract shared AI output base module from init
